Move binary deserialize internal functions to inside module
The binary tape deserializer was the only crate function that used
`object_len` and `array_len` to get the size of arguments. When jomini
was compiled without features, this would leave to warnings that these
functions were unused. The fix is to move these functions inside the
deserialize module so when deserialize is not enabled it will disable
these functions too
